Server/requirements.txt
grpcio==1.38.1
grpcio-tools==1.38.1
grpc-stubs @ git+https://github.com/MHDante/grpc-stubs.git#egg=grpc-stubs
mypy-protobuf==2.8
protobuf==3.17.3
six==1.15.0
debugpy==1.2.1
